+++ spec/unit/shoulda/matchers/active_model/validate_absence_of_matcher_spec.rb
@@ -32,7 +32,27 @@ describe Shoulda::Matchers::ActiveModel::ValidateAbsenceOfMatcher, type: :model 
             expect(validating_absence_of(:attr, {}, type: type)).
               to validate_absence_of(:attr)
           end
+
+          it_supports(
+            'ignoring_interference_by_writer',
+            tests: {
+              raise_if_not_qualified: {
+                changing_values_with: :next_value
+              },
+              accept_if_qualified_but_changing_value_does_not_interfere: {
+                changing_values_with: :next_value
+              },
+            }
+          )
+
+          define_method(:validation_matcher_scenario_args) do |*args|
+            super(*args).deep_merge(column_type: type)
+          end
         end
+      end
+
+      def validation_matcher_scenario_args
+        super.deep_merge(model_creator: :active_record)
       end
     end
 
@@ -62,6 +82,22 @@ Example did not properly validate that :attr is empty/falsy.
       it 'does not override the default message with a blank' do
         expect(active_model_validating_absence_of(:attr)).to validate_absence_of(:attr).with_message(nil)
       end
+
+      it_supports(
+        'ignoring_interference_by_writer',
+        tests: {
+          raise_if_not_qualified: {
+            changing_values_with: :upcase
+          },
+          accept_if_qualified_but_changing_value_does_not_interfere: {
+            changing_values_with: :upcase
+          },
+        }
+      )
+
+      def validation_matcher_scenario_args
+        super.deep_merge(model_creator: :active_model)
+      end
     end
 
     context 'an ActiveModel class without an absence validation' do
@@ -73,7 +109,7 @@ Example did not properly validate that :attr is empty/falsy.
         MESSAGE
 
         assertion = lambda do
-           expect(active_model_with(:attr)).to validate_absence_of(:attr)
+          expect(active_model_with(:attr)).to validate_absence_of(:attr)
         end
 
         expect(&assertion).to fail_with_message(message)
@@ -83,6 +119,22 @@ Example did not properly validate that :attr is empty/falsy.
     context 'a has_many association with an absence validation' do
       it 'requires the attribute to not be set' do
         expect(having_many(:children, absence: true)).to validate_absence_of(:children)
+      end
+
+      it_supports(
+        'ignoring_interference_by_writer',
+        tests: {
+          raise_if_not_qualified: {
+            changing_values_with: :next_value
+          },
+          accept_if_qualified_but_changing_value_does_not_interfere: {
+            changing_values_with: :next_value
+          },
+        }
+      )
+
+      def validation_matcher_scenario_args
+        super.deep_merge(model_creator: :"active_record/has_many")
       end
     end
 
@@ -97,6 +149,22 @@ Example did not properly validate that :attr is empty/falsy.
       it 'accepts' do
         model = having_and_belonging_to_many(:children, absence: true)
         expect(model).to validate_absence_of(:children)
+      end
+
+      it_supports(
+        'ignoring_interference_by_writer',
+        tests: {
+          raise_if_not_qualified: {
+            changing_values_with: :next_value
+          },
+          accept_if_qualified_but_changing_value_does_not_interfere: {
+            changing_values_with: :next_value
+          },
+        }
+      )
+
+      def validation_matcher_scenario_args
+        super.deep_merge(model_creator: :"active_record/habtm")
       end
     end
 
@@ -145,14 +213,28 @@ Parent did not properly validate that :children is empty/falsy.
       end
     end
 
-    def validating_absence_of(attr, validation_options = {}, given_column_options = {})
+    def define_model_validating_absence_of(attr, validation_options = {}, given_column_options = {})
       default_column_options = { type: :string, options: {} }
       column_options = default_column_options.merge(given_column_options)
 
-      define_model :example, attr => column_options do
-        validates_absence_of attr, validation_options
-      end.new
+      define_model :example, attr => column_options do |model|
+        model.validates_absence_of(attr, validation_options)
+
+        if block_given?
+          yield model
+        end
+      end
     end
+
+    def validating_absence_of(attr, validation_options = {}, given_column_options = {})
+      model = define_model_validating_absence_of(
+        attr,
+        validation_options,
+        given_column_options
+      )
+      model.new
+    end
+    alias_method :build_record_validating_absence_of, :validating_absence_of
 
     def active_model_with(attr, &block)
       define_active_model_class('Example', accessors: [attr], &block).new
@@ -187,6 +269,10 @@ Parent did not properly validate that :children is empty/falsy.
           validates_absence_of plural_name
         end
       end.new
+    end
+
+    def validation_matcher_scenario_args
+      super.deep_merge(matcher_name: :validate_absence_of)
     end
   end
 end