+++ spec/unit/shoulda/matchers/active_model/validate_presence_of_matcher_spec.rb
@@ -9,6 +9,35 @@ describe Shoulda::Matchers::ActiveModel::ValidatePresenceOfMatcher, type: :model
     it 'does not override the default message with a blank' do
       expect(validating_presence).to matcher.with_message(nil)
     end
+
+    it_supports(
+      'ignoring_interference_by_writer',
+      tests: {
+        raise_if_not_qualified: {
+          changing_values_with: :never_falsy
+        },
+        accept_if_qualified_but_changing_value_does_not_interfere: {
+          changing_values_with: :nil_to_blank
+        },
+        reject_if_qualified_but_changing_value_interferes: {
+          model_name: 'Example',
+          attribute_name: :attr,
+          changing_values_with: :never_falsy,
+          expected_message: <<-MESSAGE
+Example did not properly validate that :attr cannot be empty/falsy.
+  After setting :attr to ‹nil› -- which was read back as ‹"dummy value"›
+  -- the matcher expected the Example to be invalid, but it was valid
+  instead.
+
+  As indicated in the message above, :attr seems to be changing certain
+  values as they are set, and this could have something to do with why
+  this test is failing. If you've overridden the writer method for this
+  attribute, then you may need to change it to make this test pass, or
+  do something else entirely.
+          MESSAGE
+        }
+      }
+    )
   end
 
   context 'a model without a presence validation' do
@@ -37,6 +66,39 @@ Example did not properly validate that :attr cannot be empty/falsy.
     it 'does not override the default message with a blank' do
       expect(active_model_validating_presence).to matcher.with_message(nil)
     end
+
+    it_supports(
+      'ignoring_interference_by_writer',
+      tests: {
+        raise_if_not_qualified: {
+          changing_values_with: :never_falsy
+        },
+        accept_if_qualified_but_changing_value_does_not_interfere: {
+          changing_values_with: :nil_to_blank
+        },
+        reject_if_qualified_but_changing_value_interferes: {
+          model_name: 'Example',
+          attribute_name: :attr,
+          changing_values_with: :never_falsy,
+          expected_message: <<-MESSAGE
+Example did not properly validate that :attr cannot be empty/falsy.
+  After setting :attr to ‹nil› -- which was read back as ‹"dummy value"›
+  -- the matcher expected the Example to be invalid, but it was valid
+  instead.
+
+  As indicated in the message above, :attr seems to be changing certain
+  values as they are set, and this could have something to do with why
+  this test is failing. If you've overridden the writer method for this
+  attribute, then you may need to change it to make this test pass, or
+  do something else entirely.
+          MESSAGE
+        }
+      }
+    )
+
+    def model_creator
+      :active_model
+    end
   end
 
   context 'an ActiveModel class without a presence validation' do
@@ -59,6 +121,39 @@ Example did not properly validate that :attr cannot be empty/falsy.
     it 'requires the attribute to be set' do
       expect(has_many_children(presence: true)).to validate_presence_of(:children)
     end
+
+    it_supports(
+      'ignoring_interference_by_writer',
+      tests: {
+        raise_if_not_qualified: {
+          changing_values_with: :never_falsy
+        },
+        accept_if_qualified_but_changing_value_does_not_interfere: {
+          changing_values_with: :nil_to_blank
+        },
+        reject_if_qualified_but_changing_value_interferes: {
+          model_name: 'Example',
+          attribute_name: :attr,
+          changing_values_with: :never_falsy,
+          expected_message: <<-MESSAGE
+Example did not properly validate that :attr cannot be empty/falsy.
+  After setting :attr to ‹nil› -- which was read back as ‹"dummy value"›
+  -- the matcher expected the Example to be invalid, but it was valid
+  instead.
+
+  As indicated in the message above, :attr seems to be changing certain
+  values as they are set, and this could have something to do with why
+  this test is failing. If you've overridden the writer method for this
+  attribute, then you may need to change it to make this test pass, or
+  do something else entirely.
+          MESSAGE
+        }
+      }
+    )
+
+    def model_creator
+      :"active_record/has_many"
+    end
   end
 
   context 'a has_many association without a presence validation' do
@@ -69,20 +164,56 @@ Example did not properly validate that :attr cannot be empty/falsy.
   end
 
   context 'a required has_and_belongs_to_many association' do
-    before do
-      define_model :child
-      @model = define_model :parent do
-        has_and_belongs_to_many :children
-        validates_presence_of :children
-      end.new
+    it 'accepts' do
+      expect(build_record_having_and_belonging_to_many).
+        to validate_presence_of(:children)
+    end
+
+    def build_record_having_and_belonging_to_many
       create_table 'children_parents', id: false do |t|
         t.integer :child_id
         t.integer :parent_id
       end
+
+      define_model :child
+
+      define_model :parent do
+        has_and_belongs_to_many :children
+        validates_presence_of :children
+      end.new
     end
 
-    it 'accepts' do
-      expect(@model).to validate_presence_of(:children)
+    it_supports(
+      'ignoring_interference_by_writer',
+      tests: {
+        raise_if_not_qualified: {
+          changing_values_with: :never_falsy
+        },
+        accept_if_qualified_but_changing_value_does_not_interfere: {
+          changing_values_with: :nil_to_blank
+        },
+        reject_if_qualified_but_changing_value_interferes: {
+          model_name: 'Example',
+          attribute_name: :attr,
+          changing_values_with: :never_falsy,
+          expected_message: <<-MESSAGE
+Example did not properly validate that :attr cannot be empty/falsy.
+  After setting :attr to ‹nil› -- which was read back as ‹"dummy value"›
+  -- the matcher expected the Example to be invalid, but it was valid
+  instead.
+
+  As indicated in the message above, :attr seems to be changing certain
+  values as they are set, and this could have something to do with why
+  this test is failing. If you've overridden the writer method for this
+  attribute, then you may need to change it to make this test pass, or
+  do something else entirely.
+          MESSAGE
+        }
+      }
+    )
+
+    def model_creator
+      :"active_record/has_and_belongs_to_many"
     end
   end
 
@@ -204,13 +335,13 @@ raising a validation exception on failure.
     end
   end
 
-  context 'when the attribute typecasts nil to an empty array' do
-    it 'accepts' do
-      model = define_active_model_class :example do
-        attr_reader :foo
+  context 'when the attribute typecasts nil to another blank value, such as an empty array' do
+    it 'accepts (and does not raise an AttributeChangedValueError)' do
+      model = define_active_model_class :example, accessors: [:foo] do
+        validates_presence_of :foo
 
         def foo=(value)
-          @foo = Array.wrap(value)
+          super(Array.wrap(value))
         end
       end
 
@@ -250,5 +381,12 @@ raising a validation exception on failure.
     define_active_resource_class :foo, attr: :string do
       validates_presence_of :attr
     end.new
+  end
+
+  def validation_matcher_scenario_args
+    super.deep_merge(
+      matcher_name: :validate_presence_of,
+      model_creator: :active_record
+    )
   end
 end